Add compression/decompression validation on pseudorandom text data
If we are only testing handpicked sentence data, our only opportunity to catch failures is manually testing and adding testing after the fact. It would be best to additonally pass in some pseudorandom text data to pass in as a form of fuzzing/random testing.

This adds some rudimentary basic sentence like data with some random data as well as similar words for compression.